BLADERUNNER: Don't upload intangible clues to Mainframe
Change only for Restored Content mode

If KIA hacking is to be important, the player must be able to control exactly what is uploaded to Mainframe.
Intangible clues do not show up in KIA, so the player cannot control sharing them with Mainframe.
Most of them are supposed to not be shared anyway (or some are wrongly set as Intangible and will be fixed in a future bug fix).
Also in Restored Content mode, the mainframe reports (with voiced quote) if clues have been transfered,
so the player would be confused if intangible clues were transfered, since they won't appear in the KIA.