Revert "OPENGL: Fix aspect ratio bug when switching render"
This reverts commit bfceb0481350cb20974994a709086dfe34d9530a.

That commit introduced a regression causing toggling the
aspect ratio correction using the hotkey to be ignored. And
I checked that reverting the commit did not reintroduced the
bug that it was supposed to fix. However since commit a6451fff9
the window size in OpenGL mode no longer depends on whether
the Aspect Ratio correction is enabled or not.