DIRECTOR: allow setting of video casts to castNum
The game Majestic dynamically sets the castNum to video casts and expects them to play directly.
Videos were also stopped erroneously. Video stoppage is assumed when
movieRate is 0. The comment als indicated that it was probably broken
for D4.

The corresponding lingo code is, where anim is a cast label.
  set the castNum of sprite 48 to cast(anim)
  updateStage()
  set the movieRate of sprite 48 to 1
  repeat while the movieRate of sprite 48
    updateStage()
  end repeat