CINE: Use less generic file for detection
The filename `demo` was used to detect the Operation Stealth Amiga
demo. This would often produce false matches when the same filename
was present in data files for other engines, especially Mac Director.

We can use the filename `demo_os` instead. This is the file used for
the DOS demo (with a different md5).