SCUMM: Make Mac Loom practice box clickable
It turns out that the practice box (which shows note names in practice
mode) is a verb. The DOS version turns this verb on and off when needed,
but Mac Loom only sets variable 50 and 54. I have no idea why there are
two different variables, because variable 50 seems to be enough.

Anyway, this simply turns the verb on and off, and adjusts its
coordinates. It seems to work well enough.