DIRECTOR: Add non-English detection entries
Adds 14 new gameIDs and 16 new md5 entries.

This commit covers products without an English language release.
Please note that this does not necessarily indicate the country of
origin. Game descriptions for various languages are added here based
on the developers' language, whether or not we posess a copy in that
language.

This commit includes:
* 2 Danish games (+1 screen saver)
* 2 French games
* 3 Italian games
* 2 Spanish multimedia CDs
* 3 Russian demos
* 1 Brazilian catalog