SURFACESDL: Fix grabOverlay() regression
The grabOverlay() function would only copy half the width of the
overlay. This was probably more or less visible in any game that could
show a ScummVM dialog (e.g. the pause dialog in SCUMM games, or any
message box in the Hugo games).

Is crossBlit() overkill here? Probably, but it should fall back on
copyBlit() if the source and destination formats are the same.

Besides, this way perhaps I will finally get some more feedback on how
to fix the regression.