STARK: Avoid adding \0 to a String
We are still deciding on whether or not that should even be allowed, so
let's get rid of the cases where it's not actively needed. The purpose
here is to skip past unknown parts of the stream, so there's no need
storing the data in a String. Let's just document that it's ok for it to
skip past \0 as well.

This happens in my Swedish CD version of The Longest Journey, but not in
my English DVD version.