DIRECTOR: Handle FXmp mappings in loadFontMapV4
The previous way I handled the FXmp mappings overcomplicated MacText.
It also resulted in behavior different from the original Director.  With
the previous implementation, if a Win movie with Arial text was opened
with the platform set to Mac, the textFont of field 1 would return
Arial, while it would return the mapped font name Helvetica in the
original. This implementation's behavior should match the original.