SCUMM: Extended old workaround for overlapping MI2 music
While testing the Mac version of MI2, I noticed that the game would
resume the Woodtick music while Largo's theme was still playing after
the scene with the bartender. We already have a workaround for a similar
problem when he's talking to Mad Marty, and the cause seems to be the
same both times: Optimistic scripting. So delay that music transition as
well.

I've excluded the Amiga version from this workaround, because that one
cuts much of the music so it shouldn't be needed there. But if I'm
wrong, please let me know.