BLADERUNNER: fix encodings for save descriptions
If you type save names like e. g. "äöüß" in the GMM this will now show up correctly in the ingame save/load dialogs.

You still cannot type these characters ingame due to the current implementation of UIInputBox::charIsValid(char kc). I don't know why. If I comment out this function and simply return true it works just fine for me. So it is not a limitation of the font (although I am playing the English version).

Still, I haven't changed that behavior, since I am not familiar with the engine...If the only reason for charIsValid() was the wrong encoding it might make sense to limit or even remove it.