ULTIMA8: Handle NPC death a bit better
The isDead intrinsic should default to 1 if queried item is not an Actor (fixes
biohazard buzzes appearing on dead actors in No Regret).  This may fix some
subtle bugs in Ultima 8 and No Regret too as all the games return 1 for a
non-npc object.

Also when actors die in Crusader, be sure to remove them from the target list.