CGE: Fix check to skip redrawing the infoline again and again with the same text
There was a check in place to avoid redrawing the infoline when
calling the update() function with the same text multiple times.
Such calls happen when the cursor moves over an object (or menu
option) without leaving that object. The check was however
incorrect and did nothing because it was keeping the pointer
from the end of the previous text and not the start.

From the log it seems that the bug was present from the start,
and probably in the original source code. It was however fixed
in Sfinx (CGE2 engine), and this commit uses the same logic.