CONFIGURE: Do not pass --prefix to sdl2-config
sdl2-config should return the correct paths without it. Trying to determine
the path by name lookup is error-prone.

For example, if both the host and a cross-toolchain have sdl2 installed,
the host executable is typically /usr/bin/sdl2-config, and the
cross-toolchain is <triplet>-sdl2-config and
/path/to/toolchain/triplet/bin/sdl2-config.

The toolchain bin path should not be part of PATH, because this can cause
clashes when building for the host (depends on the PATH order, and if the
tool exists on both).

And even if we do add the toolchain bin path, like we currently do at least
on MXE container, if the host path comes before the toolchain then the
<triplet>-sdl2-config will be detected, but this executable *is not* in the
directory that relates to this toolchain, but in the host bin (for example,
/opt/mxe/bin), so we pass the wrong path to --prefix and sdl2 breaks.

This only worked on buildbot because the lookup for mingw was sdl-config
instead of sdl2-config, and sdl-config just doesn't exist on the system.