GUI: Display last browse path and add button for resetting it
This bumps the theme version.

The primary target is Android where a user may navigate to a location
where there is no read permission which leads to them being stuck.

This option has no effect on those native browse dialogs that do not
use the "browser_lastpath" config setting. Particularly, macOS is
not affected.