AUDIO/MIDI: Fix Miles default source volume
The default source neutral volume for the Miles MIDI driver is set at 256.
However, the default source volume was the same as that of the regular MT-32/GM
driver: 255. This would cause a small volume decrease if a game did not set
source volume.

This commit fixes this issue by setting the default source volume for the Miles
MIDI driver at 256.