ULTIMA8: Fix AdLib music
To make the Ultima 8 AdLib music sound like the original interpreter, the
MidiPlayer class has been updated to use the Miles MIDI drivers. This allows
the game's AdLib instruments to be used. Because the Audio::MidiPlayer does not
work well with the Miles drivers, I've removed the dependency on this class and
implemented the necessary functionality in the Ultima 8 MidiPlayer.