ULTIMA8: Fix SFX and speech volume control
The SFX and speech user volume controls for Ultima 8 did not work because
digital audio was played using the plain sound type instead of SFX or speech.
To distinguish between the two, a flag was added to the AudioChannel::playSample
function to indicate if the sample is a sound effect or speech.